source structure, move utility module public headers
DO NOT MERGE

Move utility public header files to include/bluetoe. This is a follow-up
to the previous commit.
