+++ bluetoe/characteristic_value.hpp
@@ -1,9 +1,9 @@
 #ifndef BLUETOE_CHARACTERISTIC_VALUE_HPP
 #define BLUETOE_CHARACTERISTIC_VALUE_HPP
 
-#include <meta_tools.hpp>
-#include <attribute.hpp>
-#include <codes.hpp>
+#include <bluetoe/meta_tools.hpp>
+#include <bluetoe/attribute.hpp>
+#include <bluetoe/codes.hpp>
 #include <meta_types.hpp>
 #include <type_traits>
 #include <climits>
