+++ bluetoe/outgoing_priority.hpp
@@ -1,7 +1,7 @@
 #ifndef BLUETOE_OUTGOING_PRIORITY_HPP
 #define BLUETOE_OUTGOING_PRIORITY_HPP
 
-#include <meta_tools.hpp>
+#include <bluetoe/meta_tools.hpp>
 
 namespace bluetoe {
