+++ bluetoe/utility/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_library(bluetoe_utility INTERFACE)
 add_library(bluetoe::utility ALIAS bluetoe_utility)
 
-target_include_directories(bluetoe_utility INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
+target_include_directories(bluetoe_utility INTERFACE include)
