+++ bluetoe/utility/include/bluetoe/client_characteristic_configuration.hpp
