+++ tests/options_tests.cpp
@@ -1,4 +1,4 @@
-#include <meta_tools.hpp>
+#include <bluetoe/meta_tools.hpp>
 #include <string>
 
 #define BOOST_TEST_MODULE
